fix type

+++ python/activator/activator.py
@@ -251,8 +251,8 @@ def parse_next_visit(http_request):
 
     # Calculate time to load knative and receive message based on time header from knative request
     _log.info("Fan out send event at %s", event['time'])
-    fan_out_knative_msg_timestamp = int(event['time'])
-    fan_out_to_prompt_time = int(time.time() - fan_out_knative_msg_timestamp)
+    fan_out_knative_msg_timestamp = float(event['time'])
+    fan_out_to_prompt_time = time.time() - fan_out_knative_msg_timestamp
     _log.info("Seconds since fan out message delivered %r", fan_out_to_prompt_time)
 
     # Message format is determined by the nextvisit-start deployment.
